+++ planarenvs/ground_robots/envs/vel_arm.py
@@ -11,32 +11,32 @@ class GroundRobotArmVelEnv(GroundRobotArmEnv):
                 "x": spaces.Box(
                     low=-self._lim_up_pos,
                     high=self._lim_up_pos,
-                    dtype=np.float64
+                    dtype=np.float32
                 ),
                 "q": spaces.Box(
                     low=-self._lim_up_arm_pos,
                     high=self._lim_up_arm_pos,
-                    dtype=np.float64,
+                    dtype=np.float32,
                 ),
                 "xdot": spaces.Box(
                     low=-self._lim_up_vel,
                     high=self._lim_up_vel,
-                    dtype=np.float64
+                    dtype=np.float32
                 ),
                 "vel": spaces.Box(
                     low=-self._lim_up_rel_vel,
                     high=self._lim_up_rel_vel,
-                    dtype=np.float64,
+                    dtype=np.float32,
                 ),
                 "qdot": spaces.Box(
                     low=-self._lim_up_arm_vel,
                     high=self._lim_up_arm_vel,
-                    dtype=np.float64,
+                    dtype=np.float32,
                 ),
             }
         )
         a = np.concatenate((self._lim_up_rel_vel, self._lim_up_arm_vel))
-        self.action_space = spaces.Box(low=-a, high=a, dtype=np.float64)
+        self.action_space = spaces.Box(low=-a, high=a, dtype=np.float32)
 
     def integrate(self):
         super().integrate()
