+++ planarenvs/planar_common/vel_env.py
@@ -7,15 +7,15 @@ class VelEnv(object):
         self.observation_space = spaces.Dict(
             {
                 "x": spaces.Box(
-                    low=-self._lim_up_pos, high=self._lim_up_pos, dtype=np.float64
+                    low=-self._lim_up_pos, high=self._lim_up_pos, dtype=np.float32
                 ),
                 "xdot": spaces.Box(
-                    low=-self._lim_up_vel, high=self._lim_up_vel, dtype=np.float64
+                    low=-self._lim_up_vel, high=self._lim_up_vel, dtype=np.float32
                 ),
             }
         )
         self.action_space = spaces.Box(
-            low=-self._lim_up_vel, high=self._lim_up_vel, dtype=np.float64
+            low=-self._lim_up_vel, high=self._lim_up_vel, dtype=np.float32
         )
 
     def integrate(self):
